requirements.txt

Django==3.2.8
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
